+++ vortex-array/src/array/bool/compute/boolean.rs
@@ -1,7 +1,5 @@
 use arrow_arith::boolean;
 use arrow_array::cast::AsArray as _;
-use arrow_array::{Array as _, BooleanArray};
-use arrow_buffer::{buffer_bin_and_not, BooleanBuffer};
 use vortex_error::VortexResult;
 
 use crate::array::BoolArray;
@@ -19,7 +17,7 @@ impl OrFn for BoolArray {
 
         let array = boolean::or(lhs, rhs)?;
 
-        null_as_false(&array)
+        Ok(Array::from_arrow(&array, false))
     }
 }
 
@@ -33,28 +31,6 @@ impl AndFn for BoolArray {
 
         let array = boolean::and(lhs, rhs)?;
 
-        null_as_false(&array)
-    }
-}
-
-/// Mask all null values of a Arrow boolean array to false
-fn null_as_false(array: &BooleanArray) -> VortexResult<Array> {
-    let inner_bool_buffer = array.values();
-
-    match array.nulls() {
-        None => Ok(Array::from_arrow(array, false)),
-        Some(nulls) => {
-            let buff = buffer_bin_and_not(
-                inner_bool_buffer.inner(),
-                inner_bool_buffer.offset(),
-                nulls.buffer(),
-                nulls.offset(),
-                inner_bool_buffer.len(),
-            );
-            let bool_buffer =
-                BooleanBuffer::new(buff, inner_bool_buffer.offset(), inner_bool_buffer.len());
-            let arr = BooleanArray::from(bool_buffer);
-            Ok(Array::from_arrow(&arr, false))
-        }
+        Ok(Array::from_arrow(&array, false))
     }
 }
