+++ vortex-datafusion/src/persistent/opener.rs
@@ -1,13 +1,16 @@
 use std::sync::Arc;
 
-use arrow_array::RecordBatch;
+use arrow_array::{Array as _, BooleanArray, RecordBatch};
 use arrow_schema::SchemaRef;
+use datafusion::arrow::buffer::{buffer_bin_and_not, BooleanBuffer};
 use datafusion::datasource::physical_plan::{FileMeta, FileOpenFuture, FileOpener};
 use datafusion_common::Result as DFResult;
 use datafusion_physical_expr::PhysicalExpr;
 use futures::{FutureExt as _, TryStreamExt};
 use object_store::ObjectStore;
-use vortex::Context;
+use vortex::array::BoolArray;
+use vortex::arrow::FromArrowArray;
+use vortex::{Array, Context, IntoArrayVariant as _};
 use vortex_error::VortexResult;
 use vortex_serde::io::ObjectStoreReadAt;
 use vortex_serde::layouts::reader::builder::VortexLayoutReaderBuilder;
@@ -64,7 +67,9 @@ impl FileOpener for VortexFileOpener {
                     async move {
                         let array = if let Some(predicate) = predicate.as_ref() {
                             let predicate_result = predicate.evaluate(&array)?;
-                            vortex::compute::filter(&array, &predicate_result)?
+
+                            let filter_array = null_as_false(&predicate_result.into_bool()?)?;
+                            vortex::compute::filter(&array, &filter_array)?
                         } else {
                             array
                         };
@@ -77,4 +82,28 @@ impl FileOpener for VortexFileOpener {
         }
         .boxed())
     }
+}
+
+/// Mask all null values of a Arrow boolean array to false
+fn null_as_false(array: &BoolArray) -> VortexResult<Array> {
+    let array = BooleanArray::from(array.boolean_buffer());
+
+    let boolean_array = match array.nulls() {
+        None => array,
+        Some(nulls) => {
+            let inner_bool_buffer = array.values();
+            let buff = buffer_bin_and_not(
+                inner_bool_buffer.inner(),
+                inner_bool_buffer.offset(),
+                nulls.buffer(),
+                nulls.offset(),
+                inner_bool_buffer.len(),
+            );
+            let bool_buffer =
+                BooleanBuffer::new(buff, inner_bool_buffer.offset(), inner_bool_buffer.len());
+            BooleanArray::from(bool_buffer)
+        }
+    };
+
+    Ok(Array::from_arrow(&boolean_array, false))
 }
